Refactor form import structure
