+++ app/forms/__init__.py
@@ -1,0 +1,3 @@
+from .error_messages import error_messages
+
+__all__ = ["error_messages"]
