+++ app/forms/field_handlers/text_area_handler.py
@@ -2,7 +2,7 @@ from functools import cached_property
 
 from wtforms import validators
 
-from app.forms.fields.max_text_area_field import MaxTextAreaField
+from app.forms.fields import MaxTextAreaField
 from app.forms.field_handlers.field_handler import FieldHandler
