+++ tests/app/forms/test_duration_form.py
@@ -1,5 +1,5 @@
 from app.forms.duration_form import get_duration_form
-from app.forms.error_messages import error_messages
+from app.forms import error_messages
 from tests.app.app_context_test_case import AppContextTestCase
