+++ tests/app/forms/test_field_factory.py
@@ -1,7 +1,7 @@
 from app.data_model.answer_store import AnswerStore
 
-from app.forms.field_factory import get_field_handler
-from app.forms.error_messages import error_messages
+from app.forms.field_handlers import get_field_handler
+from app.forms import error_messages
 from tests.app.app_context_test_case import AppContextTestCase
