+++ tests/app/forms/test_questionnaire_form.py
@@ -179,7 +179,7 @@ class TestQuestionnaireForm(
             self.assertEqual(
                 form.question_errors["date-range-question"],
                 schema.error_messages["DATE_PERIOD_TOO_LARGE"]
-                % dict(max="1 month, 20 days"),
+                % {"max": "1 month, 20 days"},
                 AnswerStore(),
             )
 
@@ -217,7 +217,7 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.question_errors["date-range-question"],
-                schema.error_messages["DATE_PERIOD_TOO_SMALL"] % dict(min="23 days"),
+                schema.error_messages["DATE_PERIOD_TOO_SMALL"] % {"min": "23 days"},
                 AnswerStore(),
             )
 
@@ -290,13 +290,13 @@ class TestQuestionnaireForm(
             self.assertEqual(
                 form.errors["date-range-from"]["year"][0],
                 schema.error_messages["SINGLE_DATE_PERIOD_TOO_EARLY"]
-                % dict(min="1 January 2017"),
+                % {"min": "1 January 2017"},
             )
 
             self.assertEqual(
                 form.errors["date-range-to"]["year"][0],
                 schema.error_messages["SINGLE_DATE_PERIOD_TOO_LATE"]
-                % dict(max="14 March 2017"),
+                % {"max": "14 March 2017"},
             )
 
     def test_date_combined_range_too_small_validation(self):
@@ -334,7 +334,7 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.question_errors["date-range-question"],
-                schema.error_messages["DATE_PERIOD_TOO_SMALL"] % dict(min="10 days"),
+                schema.error_messages["DATE_PERIOD_TOO_SMALL"] % {"min": "10 days"},
             )
 
     def test_date_combined_range_too_large_validation(self):
@@ -372,7 +372,7 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.question_errors["date-range-question"],
-                schema.error_messages["DATE_PERIOD_TOO_LARGE"] % dict(max="50 days"),
+                schema.error_messages["DATE_PERIOD_TOO_LARGE"] % {"max": "50 days"},
             )
 
     def test_date_mm_yyyy_combined_single_validation(self):
@@ -409,13 +409,13 @@ class TestQuestionnaireForm(
             self.assertEqual(
                 form.errors["date-range-from"]["year"][0],
                 schema.error_messages["SINGLE_DATE_PERIOD_TOO_EARLY"]
-                % dict(min="November 2016"),
+                % {"min": "November 2016"},
             )
 
             self.assertEqual(
                 form.errors["date-range-to"]["year"][0],
                 schema.error_messages["SINGLE_DATE_PERIOD_TOO_LATE"]
-                % dict(max="June 2017"),
+                % {"max": "June 2017"},
             )
 
     def test_date_mm_yyyy_combined_range_too_small_validation(self):
@@ -451,7 +451,7 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.question_errors["date-range-question"],
-                schema.error_messages["DATE_PERIOD_TOO_SMALL"] % dict(min="2 months"),
+                schema.error_messages["DATE_PERIOD_TOO_SMALL"] % {"min": "2 months"},
             )
 
     def test_date_mm_yyyy_combined_range_too_large_validation(self):
@@ -487,7 +487,7 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.question_errors["date-range-question"],
-                schema.error_messages["DATE_PERIOD_TOO_LARGE"] % dict(max="3 months"),
+                schema.error_messages["DATE_PERIOD_TOO_LARGE"] % {"max": "3 months"},
             )
 
     def test_date_yyyy_combined_single_validation(self):
@@ -518,13 +518,12 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.errors["date-range-from"]["year"][0],
-                schema.error_messages["SINGLE_DATE_PERIOD_TOO_EARLY"]
-                % dict(min="2015"),
+                schema.error_messages["SINGLE_DATE_PERIOD_TOO_EARLY"] % {"min": "2015"},
             )
 
             self.assertEqual(
                 form.errors["date-range-to"]["year"][0],
-                schema.error_messages["SINGLE_DATE_PERIOD_TOO_LATE"] % dict(max="2021"),
+                schema.error_messages["SINGLE_DATE_PERIOD_TOO_LATE"] % {"max": "2021"},
             )
 
     def test_date_yyyy_combined_range_too_small_validation(self):
@@ -555,7 +554,7 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.question_errors["date-range-question"],
-                schema.error_messages["DATE_PERIOD_TOO_SMALL"] % dict(min="2 years"),
+                schema.error_messages["DATE_PERIOD_TOO_SMALL"] % {"min": "2 years"},
             )
 
     def test_date_yyyy_combined_range_too_large_validation(self):
@@ -586,7 +585,7 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.question_errors["date-range-question"],
-                schema.error_messages["DATE_PERIOD_TOO_LARGE"] % dict(max="3 years"),
+                schema.error_messages["DATE_PERIOD_TOO_LARGE"] % {"max": "3 years"},
             )
 
     def test_bespoke_message_for_date_validation_range(self):
@@ -922,7 +921,7 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.question_errors["breakdown-question"],
-                schema.error_messages["TOTAL_SUM_NOT_EQUALS"] % dict(total="10"),
+                schema.error_messages["TOTAL_SUM_NOT_EQUALS"] % {"total": "10"},
                 AnswerStore(),
             )
 
@@ -997,7 +996,7 @@ class TestQuestionnaireForm(
             self.assertEqual(form.data, expected_form_data)
             self.assertEqual(
                 form.question_errors["breakdown-question"],
-                schema.error_messages["TOTAL_SUM_NOT_EQUALS"] % dict(total="10"),
+                schema.error_messages["TOTAL_SUM_NOT_EQUALS"] % {"total": "10"},
                 AnswerStore(),
             )
 
@@ -1050,7 +1049,7 @@ class TestQuestionnaireForm(
 
             self.assertEqual(
                 form.question_errors["breakdown-question"],
-                schema.error_messages["TOTAL_SUM_NOT_EQUALS"] % dict(total="10"),
+                schema.error_messages["TOTAL_SUM_NOT_EQUALS"] % {"total": "10"},
             )
 
     def test_generate_form_with_title_and_no_answer_label(self):
@@ -1207,7 +1206,7 @@ class TestQuestionnaireForm(
             form.validate()
             answer_errors = form.answer_errors("set-minimum")
             self.assertIn(
-                schema.error_messages["NUMBER_TOO_SMALL"] % dict(min="0"), answer_errors
+                schema.error_messages["NUMBER_TOO_SMALL"] % {"min": "0"}, answer_errors
             )
 
     def test_mandatory_mutually_exclusive_question_raises_error_when_not_answered(self):
@@ -1233,7 +1232,7 @@ class TestQuestionnaireForm(
             )
 
     def test_mandatory_mutually_exclusive_question_raises_error_with_question_text(
-        self
+        self,
     ):
         with self.app_request_context():
             schema = load_schema_from_name("test_question_title_in_error")
