+++ tests/app/forms/validation/test_date_check_validator.py
@@ -2,7 +2,7 @@ import unittest
 from unittest.mock import Mock
 from wtforms.validators import StopValidation
 
-from app.forms.error_messages import error_messages
+from app.forms import error_messages
 from app.forms.validators import DateCheck
