+++ tests/app/forms/validation/test_single_date_range_validator.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock
 from wtforms.validators import ValidationError
 
 from app.questionnaire.rules import convert_to_datetime
-from app.forms.error_messages import error_messages
+from app.forms import error_messages
 from app.forms.validators import SingleDatePeriodCheck
 from tests.app.app_context_test_case import AppContextTestCase
 
@@ -23,7 +23,7 @@ class TestDateRangeValidator(AppContextTestCase):
 
             self.assertEqual(
                 error_messages["SINGLE_DATE_PERIOD_TOO_EARLY"]
-                % dict(min="30 March 2016"),
+                % {"min": "30 March 2016"},
                 str(ite.exception),
             )
 
@@ -41,8 +41,7 @@ class TestDateRangeValidator(AppContextTestCase):
                 validator(mock_form, mock_field)
 
             self.assertEqual(
-                error_messages["SINGLE_DATE_PERIOD_TOO_LATE"]
-                % dict(max="1 April 2016"),
+                error_messages["SINGLE_DATE_PERIOD_TOO_LATE"] % {"max": "1 April 2016"},
                 str(ite.exception),
             )
